✨ (Home.test.jsx): Add tests for the Home component to ensure proper rendering and functionality 📝 (jest.config.ts): Configure Jest settings for testing environment and coverage reporting 🔧 (jest.setup.ts): Add jest setup file to configure testing environment with '@testing-library/jest-dom' 🔧 (package.json): Add Jest as a test script and test:watch script for running tests in watch mode The changes were made to enhance the testing capabilities of the application by adding test files for the Home component, configuring Jest settings, setting up Jest environment, and adding Jest as a test script in package.json for easier test execution.
✨ (page.tsx): Create a new HomePage component to handle the logic for displaying articles on the home page
📝 (Home.tsx): Move the existing Home component logic to a new file to separate concerns and improve code organization
The changes were made to separate the logic for displaying articles on the home page into a dedicated component called HomePage. This separation of concerns improves code organization and maintainability by isolating the specific functionality related to handling articles. Additionally, moving the existing Home component logic to a new file named Home.tsx helps in better structuring the codebase and keeping related functionalities together.
