+++ package.json
@@ -7,7 +7,9 @@
     "build": "next build",
     "postbuild": "node scripts/sitemap.mjs",
     "start": "next start",
-    "lint": "next lint"
+    "lint": "next lint",
+    "test": "jest",
+    "test:watch": "jest --watch"
   },
   "browserslist": "defaults, not ie <= 11",
   "dependencies": {
@@ -43,13 +45,19 @@
     "reading-time-estimator": "^1.11.0",
     "remark-gfm": "^4.0.0",
     "tailwindcss": "^3.4.1",
+    "ts-jest": "^29.1.2",
+    "ts-node": "^10.9.2",
     "typescript": "^5.3.3",
     "usehooks-ts": "^3.1.0"
   },
   "devDependencies": {
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^15.0.2",
     "@types/react-instantsearch-dom": "^6.12.3",
     "eslint": "^8.56.0",
     "eslint-config-next": "^14.2.1",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "prettier": "^3.1.1",
     "prettier-plugin-tailwindcss": "^0.5.11",
     "sharp": "0.33.1"
