+++ src/components/home/index.tsx
@@ -1,3 +1,4 @@
+export { default as Home } from './Home'
 export { default as Messenger } from './Messenger'
 export { default as Newsletter } from './Newsletter'
 export { default as Resume } from './Resume'
